+++ patches/src/main/kotlin/app/morphe/patches/music/misc/litho/filter/LithoFilterPatch.kt
@@ -7,11 +7,11 @@
 
 @file:Suppress("SpellCheckingInspection")
 
-package app.morphe.patches.music.misc.litho
+package app.morphe.patches.music.misc.litho.filter
 
 import app.morphe.patches.music.misc.extension.sharedExtensionPatch
 import app.morphe.patches.music.misc.litho.context.conversionContextPatch
-import app.morphe.patches.shared.misc.litho.sharedLithoFilterPatch
+import app.morphe.patches.shared.misc.litho.filter.sharedLithoFilterPatch
 
 val lithoFilterPatch = sharedLithoFilterPatch(
     // Supported YT Music versions always use the native Upb encode path.
